+++ test/protractor/spec/lforms_formdata.spec.js
@@ -213,7 +213,7 @@ describe('get form data', function() {
       // console.log(formData);
       expect(formData.items.length).toBe(1);
       expect(formData.items[0].question).toBe("With data type CNE");
-      expect(formData.items[0].value).toEqual({code:"c2",other:null,text:"Answer 2"});
+      expect(formData.items[0].value).toEqual({code:"c2",other:null,text:"Answer 2", "_displayText": "Answer 2"});
     });
   });
 });
@@ -254,7 +254,8 @@ describe('defaultAnswer', function() {
     expect(element(by.id(radioQCode+'G')).isSelected()).toBe(false);
     element(by.id(radioQCode+'R')).evaluate('item.value').then(function(val) {
       expect(val.code).toEqual('B');
-      expect(val.text).toEqual('ii. Blue');
+      expect(val.text).toEqual('Blue');
+      expect(val._displayText).toEqual('ii. Blue');
     });
 
     // Check a radio button question whose answers do not have labels
@@ -274,7 +275,8 @@ describe('defaultAnswer', function() {
     expect(element(by.id(cbQCode+'G')).isSelected()).toBe(false);
     element(by.id(cbQCode+'R')).evaluate('item.value').then(function(val) {
       expect(val[0].code).toEqual('B');
-      expect(val[0].text).toEqual('ii. Blue');
+      expect(val[0].text).toEqual('Blue');
+      expect(val[0]._displayText).toEqual('ii. Blue');
     });
 
     // Check a multi-select list
@@ -282,8 +284,9 @@ describe('defaultAnswer', function() {
     var multiSel = element(by.id(multiSelID));
     multiSel.evaluate('item.value').then(function(value) {
       expect(value.length).toBe(1);
-      expect(value[0].text).toEqual('ii. Blue');
+      expect(value[0].text).toEqual('Blue');
       expect(value[0].code).toEqual('B');
+      expect(value[0]._displayText).toEqual('ii. Blue');
     });
     expect(multiSel.getAttribute('value')).toEqual('');
     var escapedID = multiSelID.replace( /\//g, "\\\\/" );